feature: add components to render the links and the footer for the sidebar and remove hardcoded components from desktopbar
